improve settings UI
